+++ infra/aws-cdk/lib/platform-stack/index.ts
@@ -21,40 +21,26 @@ export class FormsPlatformStack extends cdk.Stack {
       description: 'The environment for the stack (e.g., dev, prod)',
       allowedPattern: '.+',
     });
-    const dockerImagePath = new cdk.CfnParameter(this, 'imageUri', {
+    const repositoryArn = new cdk.CfnParameter(this, 'repositoryArn', {
       type: 'String',
-      description: 'The Docker image url for the App Runner service',
+      description: 'The ARN for the ECR repository',
       allowedPattern: '.+',
-      default: 'public.ecr.aws/aws-containers/hello-app-runner:latest',
+    });
+    const repositoryName = new cdk.CfnParameter(this, 'repositoryName', {
+      type: 'String',
+      description: 'The ECR repository name for the deploy image',
+      allowedPattern: '.+',
+    });
+    const tagOrDigest = new cdk.CfnParameter(this, 'tagOrDigest', {
+      type: 'String',
+      description: 'The ECR image tag or digest to deploy',
+      allowedPattern: '.+',
     });
 
     // Networking configuration
     const vpc = new ec2.Vpc(this, `${id}-vpc`, {
       maxAzs: 2,
     });
-
-    // Get the default security group
-    const defaultSecurityGroup = ec2.SecurityGroup.fromSecurityGroupId(
-      this,
-      'DefaultSecurityGroup',
-      vpc.vpcDefaultSecurityGroup
-    );
-
-    // Explicitly remove inbound and outbound rules.
-    // This is an alternative to the cdk.json setting:
-    //  "@aws-cdk/aws-ec2:restrictDefaultSecurityGroup": true
-    defaultSecurityGroup.addEgressRule(
-      ec2.Peer.anyIpv4(),
-      ec2.Port.allTraffic(),
-      'Remove all egress traffic',
-      false
-    );
-    defaultSecurityGroup.addIngressRule(
-      ec2.Peer.anyIpv4(),
-      ec2.Port.allTraffic(),
-      'Remove all ingress traffic',
-      false
-    );
 
     // Security group for RDS
     const rdsSecurityGroup = new ec2.SecurityGroup(this, `${id}-rds-sg`, {
@@ -74,14 +60,27 @@ export class FormsPlatformStack extends cdk.Stack {
       }
     );
 
-    // Allow App Runner security group to access RDS security group
-    /*
-    rdsSecurityGroup.addIngressRule(
-      appRunnerSecurityGroup,
-      ec2.Port.tcp(5432),
-      'Allow postgres access from App Runner'
+    const loginGovSandboxIpAddresses = [
+      '108.156.107.56',
+      '108.156.107.41',
+      '108.156.107.98',
+      '108.156.107.80',
+    ];
+    loginGovSandboxIpAddresses.forEach(ip =>
+      appRunnerSecurityGroup.addEgressRule(
+        ec2.Peer.anyIpv4(),
+        ec2.Port.tcp(443),
+        'Allow outbound HTTPS traffic to idp.int.identitysandbox.gov'
+      )
     );
-    */
+
+    new ec2.InterfaceVpcEndpoint(this, 'SecretsManagerVPCEndpoint', {
+      vpc,
+      service: ec2.InterfaceVpcEndpointAwsService.SECRETS_MANAGER,
+      subnets: { subnetType: ec2.SubnetType.PUBLIC },
+      securityGroups: [appRunnerSecurityGroup],
+    });
+
     new ec2.CfnSecurityGroupIngress(this, `${id}-apprunner-rds-ingress`, {
       groupId: rdsSecurityGroup.securityGroupId,
       sourceSecurityGroupId: appRunnerSecurityGroup.securityGroupId,
@@ -125,43 +124,30 @@ export class FormsPlatformStack extends cdk.Stack {
     dbSecret.grantRead(appRunnerRole);
 
     const appRunnerService = new apprunner.Service(this, `${id}-app-runner`, {
-      source: dockerImagePath.valueAsString.endsWith('.amazonaws.com')
-        ? apprunner.Source.fromEcr({
-            imageConfiguration: {
-              port: 4321,
-              environmentVariables: {
-                DB_HOST: rdsInstance.dbInstanceEndpointAddress,
-                DB_PORT: rdsInstance.dbInstanceEndpointPort,
-                DB_NAME: 'postgres',
-              },
-              environmentSecrets: {
-                DB_SECRET_ARN: apprunner.Secret.fromSecretsManager(dbSecret),
-              },
-            },
-            repository: ecr.Repository.fromRepositoryAttributes(
-              this,
-              `${id}-repo`,
-              {
-                repositoryName: `forms-platform-${environment.valueAsString}`,
-                repositoryArn: `arn:aws:ecr:us-east-2:001907687576:repository/forms-platform-${environment.valueAsString}`,
-              }
-            ),
-            tagOrDigest: 'latest',
-          })
-        : apprunner.Source.fromEcrPublic({
-            imageConfiguration: {
-              port: 4321,
-              environmentVariables: {
-                DB_HOST: rdsInstance.dbInstanceEndpointAddress,
-                DB_PORT: rdsInstance.dbInstanceEndpointPort,
-                DB_NAME: 'postgres',
-              },
-              environmentSecrets: {
-                DB_SECRET_ARN: apprunner.Secret.fromSecretsManager(dbSecret),
-              },
-            },
-            imageIdentifier: dockerImagePath.valueAsString,
-          }),
+      autoDeploymentsEnabled: true,
+      source: apprunner.Source.fromEcr({
+        imageConfiguration: {
+          port: 4321,
+          environmentVariables: {
+            DB_HOST: rdsInstance.dbInstanceEndpointAddress,
+            DB_PORT: rdsInstance.dbInstanceEndpointPort,
+            DB_NAME: 'postgres',
+          },
+          environmentSecrets: {
+            DB_SECRET: apprunner.Secret.fromSecretsManager(dbSecret),
+            //DB_SECRET_ARN: dbSecret,
+          },
+        },
+        repository: ecr.Repository.fromRepositoryAttributes(
+          this,
+          `${id}-repo`,
+          {
+            repositoryName: repositoryName.valueAsString,
+            repositoryArn: repositoryArn.valueAsString,
+          }
+        ),
+        tagOrDigest: tagOrDigest.valueAsString,
+      }),
       healthCheck: apprunner.HealthCheck.http({
         healthyThreshold: 5,
         interval: Duration.seconds(10),
@@ -187,26 +173,10 @@ export class FormsPlatformStack extends cdk.Stack {
       }),
     });
 
-    // Create an ECR repository for the App Runner service's images.
-    /*
-    const ecrRepository = new ecr.Repository(this, `${id}-ecr-repository`, {
-      repositoryName: `forms-platform-${environment.valueAsString}`,
-    });
-    ecrRepository.grantPull(appRunnerService);
-    */
-
-    // Export a publicly-accessible URL for the App Runner service≥
+    // Export a publicly-accessible URL for the App Runner service.
     new cdk.CfnOutput(this, 'FormsPlatformUrl', {
       value: appRunnerService.serviceUrl,
       description: 'URL for the Forms Platform',
     });
-
-    // Export a publicly-accessible URL for the App Runner service≥
-    /*
-    new cdk.CfnOutput(this, 'FormsPlatformEcrUrl', {
-      value: ecrRepository.repositoryUri,
-      description: 'ECR repository URI for the Forms Platform',
-    });
-    */
   }
 }