+++ infra/aws-cdk/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @gsa-tts/forms-infra-aws-cdk
 
+## 0.1.12
+
+### Patch Changes
+
+- CDK work
+- Updated dependencies
+  - @gsa-tts/forms-infra-core@0.1.4
+
+## 0.1.11
+
+### Patch Changes
+
+- VPC endpoint for AWS Secrets manager
+- Updated dependencies
+  - @gsa-tts/forms-infra-core@0.1.3
+
+## 0.1.10
+
+### Patch Changes
+
+- Include node types
+
+## 0.1.9
+
+### Patch Changes
+
+- Use standard cdk.out output dir
+
+## 0.1.8
+
+### Patch Changes
+
+- Include lib directory
+
+## 0.1.7
+
+### Patch Changes
+
+- Include tsconfig and src in package
+
+## 0.1.6
+
+### Patch Changes
+
+- Include entire bin directory in package
+
+## 0.1.5
+
+### Patch Changes
+
+- Include CDK config in package
+
+## 0.1.4
+
+### Patch Changes
+
+- Deploy FormsPlatformStack with appropriate parameters
+
+## 0.1.3
+
+### Patch Changes
+
+- 697154e: Create AWS CodePipeline CDK stack
+
 ## 0.1.2
 
 ### Patch Changes